Refactor of configurations
We now use parameters in the run_config.yaml to specify settings of the dataset. Makes it easier to play around with different settings. We give a template in res/configs/run_config_template.yaml and the settings used in the baseline in res/configs/baseline_config.yaml. The baseline_config.yaml is only missing the private api_key and what wandb-project to log into.